+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+1.5.4
+=====
+
+*   (improvement) `closest()` now properly checks the element itself. If you want the old behavior, use `closest(element.parentElement)` instead.
+*   (improvement) You can now call `closest()` with `null` as element.
+*   (bug) Fix `delegate` not actually registering the event listener.
+*   (improvement) Bump all dependencies.
+
+
 1.5.3
 =====